[CIRCT] Move parallelTransformReduce to common util header
